Experiment complete - working
description:
There is a scene with a plane and a cube on the plane.
There is a color change on the cube if you hover it.
If you click the cube the camera changes positions (animated), and a popup for extra content opens up.

bugs:
clicking fast multiple times in a row makes the extra content not disappear